Move the backend test data to 'testdata' package
